A new script!
Can be run as part of the build or on its own.

Specify a database (using -database) and this script will try to find spurious
short ORFs below a certain length threshold (defaults to 100 amino acids but can
change this by using -cutoff).

It will look for associated RNAi, PCR product and cDNA information to classify
each short gene (those with no cDNAs, RNAi experiments that give 'Wild type'
phenotypes, and PCR products that don't amplify are considered to be the worst!).

If -build is specified it will assume that you are running as part of the build
and a) will write output file to /wormsrv2/autoace/CHECKS (rather than the default
/wormsrv2/tmp) and b) will email log file

I intend to plug this into the build and then get the make_website scripts pick
this info up and put it (somehow) on the page with the other curation targets.

My script classifies genes into six categories (1 = the worst, 6 the best) and so
the webpage can organise the information to reflect this.

Keith

P.S. My script takes no account of briggsae information (from Waba) or any protein
or motif homologies

P.P.S. This all stems from my belief that many short ORFs are probably spurious,
or possibly can be merged into other gene models.