CDS_class
Some final fixes which are just reverting back to former tag names.  acedb doesn't
like the tag 'CDS' to appear in Smapped classes because it gets confused that it might
be the 'CDS' tag from the ?CDS class.

So:

In the ?Sequence class 'CDS' reverts to 'CDS_child'
In the ?Allele and ?RNAi class, 'CDS' reverts to 'Predicted_gene'

That's all

Keith