another check point to do this:
ERROR 35: Allele ju255 is linked to unc-71 max-3 and Y37D8A.13, but the locus does not have seq. info
ERROR 35: Allele nr2027 is linked to ikb-1 and C04F12.3, but the locus does not have seq. info
ERROR 35: Allele ok310 is linked to unc-57 and T04D1.3, but the locus does not have seq. info
ERROR 35: Allele ok425 is linked to vav-1 and C35B8.2, but the locus does not have seq. info
ERROR 35: Allele oxP1 is linked to oxP1 and ZK1236.3, but the locus does not have seq. info
ERROR 35: Allele oxP2 is linked to oxP2 and T19B10.5, but the locus does not have seq. info
ERROR 35: Allele pk22 is linked to pes-9 and R11H6.1, but the locus does not have seq. info
ERROR 35: Allele t1623 is linked to rot-2 and C36E8.5, but the locus does not have seq. info