Fixed a minor bug left over from my ?CDS script changes...
When the ?Database_field component of the DB_info line was being set for ?CDS objects, it was
defaulting to 'TrEMBL_ID' or 'TrEMBL_AC' even if it was a SwissProt protein.  Now fixed.  Don't
know why I hard-coded this in the first place.  Don't think there would have been any side-effects
of this though.

Keith