I can't believe that this works!
So far it takes a target database, finds all curated CDSs (which are c. elegans) and that don't have a locus connection.

It then generated ?Gene objects for all of these.  Multiple isoforms of the
same gene get connected to the same Gene object.

Still needs work to handle Pseudogene and Transcript classes and also
needs to handle C. briggsae species and then (finally) be able to work
through assigning Genes to history objects.

You also need to know what the highest available gene ID is.  This is currently
hard-coded but should be a command line option.

Keith