Penultimate set of changes for WS126:
Remove UNIQUE qualifier for 'Corresponding_transcript' tag in CDS class.  This used to be a 1-2-1
relationship with Transcript objects, but Anthony's new script will allow one CDS to have multiple
transcripts to capture variation in the 5' and 3' UTR regions.

At the same time I found that we still have a redundant 'Corresponding_transcript' tag in the
Sequence class which I have now removed.