Final model change for WS126
Two related steps:

1) misc_RNA tag in Transcript class changed to ncRNA...this means all RNA types use SOFA compliant names for their tags

2) This 'Transcript' subtree of RNA types in ?Transcript is copied and moved to ?Sequence under the existing RNA tag to give:

     RNA UNIQUE mRNA UNIQUE Processed_mRNA
                            Unprocessed_mRNA
                tRNA Type UNIQUE Text  // new for ck1 [030926] krb
		     Anticodon UNIQUE Text  //
                rRNA UNIQUE Text
		snRNA UNIQUE Text
                snoRNA UNIQUE Text // [030102 krb]
                scRNA UNIQUE Text
                stRNA UNIQUE Text         // [010723 dl]
                miRNA UNIQUE Text         // [020306 kj]
                ncRNA UNIQUE Text         // for everything else that doesn't have a known type


So we can now tag sequences corresponding to RNA genes in EMBL/GenBank entries.  This will make it much easier to query
the database to get 'all mRNA' entries etc.

THis is sort of coming full circle as these tags all used to be in the ?Sequence model but were removed to go into
?Transcript