still some improvements to geneace_check:
(1) added (a) following ERROR or UPDT, eg, to indicate ace output to directly fix problems in geneace, as some errors can only be looked at by hand and not automated this way

(2) added checking gene id with same Other_name and CGC_name

(3) for option -ps or -pseudo, it sends out a list of candidate pseudo genetic markers for Jonathan and cgc with the subject line "Pseudo genetic marker(s) to approve for WSxxx", where xxx is the version of the next build. This gives CGC 2 weeks time to approve the list and also copes with the 2 weeks cycle of WB release.

(4) added also one check to look for gene id that does not have a map positin nor interpolated_map_position


All that I can do for geneace_check for now, and on the last day of being a Wormbaser or BETTER, the start of a BEER BREWER!!!