Big change to how wormpep files are updated.
I'm in the process of bringing the web wormpep files which were just under
/htdocs/Projects/C_elegans/wormpep/ into each release directory for wormbase,
e.g.
/htdocs/Projects/C_elegans/WORMBASE/WSXXX/

This means that we keep some history of each wormpep release and tie it more
closely to each new wormbase release.  This also means that we can keep a distinction
between live and dev releases and make sure that the wormpep webpages stay in sync
with the wormbase webpages.

I've had to make lots of changes to the web pages layout to do this (currently
just on the dev site) and I still need to make one more change to update_live_site
to take this all into account.

I expect something will go wrong!  Some external web links will need to be updated
to /htdocs/Projects/C_elegans/WSXXX/wormpep.html (the new parent page for wormbase
releases)

Keith