This may fix a long standing problem where we have lost the path to the tace binary.
If a script calls a subroutine in WormBase multiple times, and that subroutine has something
like:
my $tace = glob(some path to tace);

then in this context glob acts as an iterator.  The first use of the subroutine sets
tace correctly, but the second call forces glob to return the next value in its
list, and this will now be empty as there was only one value to begin with.

The solution is to either use the return values of glob in a list (rather than scalar)
context or to set the definition of $tace outside of subroutines so that it only
gets called once per use of WormBase.pm as a whole.  This is what we have done.