Yet another common_dat file:
worm_gene2class.dat

'K02E10.5' => 'CDS',
'F09E10.11b' => 'Transcript',
'F21F8.8' => 'Pseudogene',

This might seem a little strange but if you are dealing with a worm_genes
object you might not always know what class it is, and rather than
use AcePerl to look for CDS, Pseudogene, or Transcript tags I thought
it quicker to make a dat file.

This will be used by map_Alleles.pl because the Feature_mapper module
looks for overlaps between Transcripts and CDSs but doesn't distinguish
in its output.  So it might return 'F98E10.11b' as something that
overlaps allele ok190 but you won't know that this is transcript
and not a CDS.

map_Alleles.pl currently assumes that *everything* is a CDS which is
wrong, and it currently makes acefile connections to CDSs.  I'll
now make map_Allels.pl check the class of the overlapping CDS/Transcript
and adjust the acefile output accordingly.

Keith