Lots of changes!
1) GetInterPro_motifs.pl no longer runs daily (at 2:00 pm) as part of a cronjob and no therefore
no longer writes to /wormsrv2/wormbase/misc_dynamic.  It now runs as part of the build (called
by autoace_minder.pl with the -buildpep option and writes to /wormsrv2/autoace/acefiles.  This
ace file used to be loaded because it was listed in autoace_config.  The script now has a -load
option which uses autoace_minder.pl -load to load the file.  I now need to change autoace_config

2) make_Interpro2GO_mapping.pl used to be called by getProteinID (in turn called by autoace_minder.pl)
and also wrot eto the misc_dynamic directory.  This script is no longer called by getProteinID and
is now called by autoace_minder.pl (again as part of -buildpep).  This now gets run straight after
the previous script and hopefully reduces a tier of complexity where we had a script calling a script
calling a script

3) getProteinID is now getProteinID.pl and is changed to accommodate the above details.  I also added
a -load option so that the script can now load the acefile it produces (previously an extra part of
autoace_minder.pl -buildpep loaded the acefile).

4) autoace_minder.pl -buildpep now does all of the above.  This might not be the best long term place
for all of these steps but for now it is maybe a little tidier

So we are another step closer to having acefiles made by the build stored in one place