Previously, when using Anthony's log file module, it would only ever use the script name as the subject line of the email that sent the log report.
In many cases this is fine, but in a lot of our build scripts I have
started adding things like

"BUILD REPORT: map_Alleles.pl 4 ERRORs"

The 'BUILD REPORT' part I have found useful for quickly seeing which emails
are logs from the build and also it has been useful to flag errors when
they happen.

So I've amended Anthony Log_files.pm module to optionally pass an extra
argument when sending the mail, and that argument is a text string for the
subject line.  The first argument remains the email address(es) of the
intended recipients, i.e.


$log->mail("$maintainers","BUILD REPORT: map_Alleles.pl $error_count ERRORS!");