This is an annotation GUI tool for creating history objects at the touch of a button.
A perl Tk interface presents to the user a simple box with a space to enter a CDS name and a button to make a history object.  The input to the acedb database is done immediately using 'xremote' and is visible after recalculating the fmap.  The target database is determined by the operating system as the last one to have been opened using the terminal that this script is run from.

To ensure that the correct database is being used a shell script should be used to launch both simultaneously.

A source database is hardcoded ( at the moment ) and this is used to extract the relevant info needed to make a history ie source exons, gene_id etc.

Some error checking is done so that history objects cant be created for non-existant CDSs.

This can be used with care until more error checking is in place.

A database version parameter must be passed to be appended to the history object name. eg

history_maker.pl 132

will make histories like AH6.4:wp132