Added Position_Matrix class proposed by Erich
This is intended to be a general data model for known or
inferred binding sites of transcription factors to DNA, accomodating
either a straight position-frequency matrix (PFM), or a
position-specific weight matrix (PSFM, which requires a background
sequence model).  It also provides a way to link specific sites in
the genome with PWM/PSFMs (and this linkage can be many-to-one, so
that a given site can have overlapping or similar PWM/PSFMs
associated with it).  It does not specifically try to associate
genes with PWM/PSFMs, but that could be easily added to the model if
people wanted it.

--------------------------------------

Replaced the searchable ?Text with plain old Text in Remarks in
?Oligo_set, ?Microarray_results and #Microarray_data

for much more efficient loading.