There were 2 sets of changes to the Strain model.
The Method model was brought back in-line with the Core ACeDB models.wrm file with the preservation
of the Gapped tag. This allows for backwards compatibility for older binaries and gives a cleaner path to move forward. The majority of the changes are //comments but the Fmap alignment tags for Allow_misalign was moved with the addition of some new sub-tags.

Please read in a new version of the misc_autoace_methods.ace file into your database.

Additional information can be found here: http://wiki.wormbase.org/index.php/WS224_Models.wrm