WS226 models v 1.2
The models file has been re-tagged in line with the new release schedule.

Two changes to the ?Strain class.

Introduction of an Other_name tag to hold old names associated with a species,
(pre-classification names) etc.

Removal of an XREF to protein.

Added ?Strain/?Species to a minimal set of classes to help
the new website architecture

    * ?Clone
    * ?Condition
    * ?Expr_profile
    * ?Expression_cluster
    * ?Feature_data
    * ?Microarray_experiment
    * ?PCR_product
    * ?SAGE_experiment
    * ?Transcription_factor

Other suggested classes welcome for WS227

More infromation and a human readable diff can be found here:
http://wiki.wormbase.org/index.php/WS226_Models.wrm