MAJOR UPDATE to accommodate UniProt Gene-centric Reference Proteome (GCRP) cross-references.
Thistequired changing format of table dumped from EBI ORACLE databases
(with consequent knock-on effects for some scripts), and the splitting
up of CDS and protein xrefs into two scripts (the latter being done with
a new script). This is mainly so so that non-WormBase annotations
(specificaly those on the C. elegans mitochondrion) are dealt with
correctly.