Check specified referene allele against genome sequence, and bug fix
Some entries in the varaiation JSON files have reference alleles
that don't match up with the specified genomic coordinates.  Many
(but not all) of these appear to correspond to variants annotated
against the reverse strand.  However, we cannot just reverse
complement the ref and alt alleles as we cannot validate the alt
allele, so could be replacing one error with another.  These
entries are therefore skipped.