Use checksums to avoid unnecessary analyses
Checksums of VEP input files are stored in a file.  When the
script is run, it downloads the files from AGR, calculates the
md5 checksums for the files, and compares it against those stored
in the file.  Analyses are then only run if the input files have
changed.  The checksums file is only updated if the script is not
run in test mode and all stages are run - this is to ensure the
checksums stored reflect those used for input when generating the
files at AGR.