Replace variation ID with HGVSg ID in phenotypic VEP output
The JSON to VCF conversion script that is now used to generate
VCF files from the MOD-submitted variation JSON files uses the
alleleId attribute as the variation ID.  Previously, the AGR-
generated VCFs used the HGVSg ID.  This change causes the loader
to break.

The VEP output contains the HGVSg ID in the Extras column, so we
can parse the file to replace the variation ID with the HGVSg ID.
We also convert the HGVSg ID in the extras column to a version with
the chromosome name rather than the RefSeq accession as this is
consistent with VEP files used in previous releases, and this is
stored as a synonym.

Also reimplemented submission validation check for VEP files - this
was erroneously removed as the VEP files weren't returning JSON
upon submission, but this turned out to be due to an incorrectly
set $TOKEN environment variable rather than due to the submitted
file not being a JSON file.