Rewrite to deal with non-standard VCF entries
Variations with insertions of unknown sequence need to be included.
These are represented by a flanking base and a '.' (e.g. 'C.').
This required a change of logic to use the 'type' field in the JSON,
which contains a SO term, to identify the type of variation instead
of relying on the genomic and variant sequences to determine type.

Added method to retrieve chromosome name from RefSeq chromosome
accession as chromosome is not a required field in the JSON schema.