Add clause for empty variant sequence fields
There are cases where the inserted sequence is unknown.  Normally
the JSON entry does not contain a genomicVariantSequence field,
but in some cases there is an empty string.  This was causing the
variant allele to be represented in the VCF as just the flanking
base without the dot indicating that there is an insertion sequence
of unknown composition (e.g. C instead of C.).