Replace variation IDs with HGVSg
HGVSg IDs are required for the variation IDs in the VEP output
as this is where the loader gets the HGVSg ID from (the HGVSg
generated by the VEP in the final column is loaded as a synonym).
The VEP also does not generate HGVSg IDs for variants with non-ACGT
bases in the alt allele, but we need these included.