Change in HGVSg ID logic
The HGVSg ID is required in the first column.  This was previously
parsed from the HGVSg attribute in the extras column.  This caused
isses as not all VCF entries have VEP consequences and HGVSg IDs.

HGVSg IDs are now in the VCF input files.  The HGVSg IDs in the ID
and Extras columns are compared and a warning is logged if they do
not match.  The RefSeq chromosome accession part of the HGVSg ID
in the Extras column is still switched for the chromosome number -
this is no longer a true HGVSg ID, but it is used as a synonym at
the Alliance and this is done for consistency.