Fixed chromomsome prefix bug and implemented VCF-only mode
The chromosome sequence dictionary was stripping the CHROMOSOME_
prefix from the IDs in the FASTA file, but the GFF file contains
these prefixes at the stage the script is run (stripped later
for final processed GFFs).

A VCF-only mode was implemented to allow the script to be run to
generated VCF files without altering the GFF files, and to allow
the final processed (and gzipped) GFF file to be used.