Add option for using external GFF for human
We were previusly having issues with the human GFF supplied by
RGD.  To avoid delays, a GFF was created by combining the EnsEMBL
and NCBI GFF files.  This is now no longer required, as the RGD-
supplied GFF has been fixed, but we want to keep the use of the
external GFF as an option for future use.