Expand IUPAC alternative alleles for VEP pipeline
The VEP doesn't recognise IUPAC ambiguity codes, so for variations
with these codes that lie in coding regions it cannot determine the
coding sequence and a 'coding_sequence_variation' consequence is
returned with 'MODIFIER' impact.

This update subsitutes in the possible bases at the points of the
IUPAC ambiguity codes in order to give all permutations.  For each
affected transcript, the most severe consequence is stored in
AceDB and propagated to the annotations in our GFF and VCF files.

Further changes may be required to the codebase and to the model
if we want to be able to store all possible consequences.