Update BLAST ace files tests to use files from previous build
Table-Maker queries of previous build take a long time for these
tests.  This should not be necessary as the ace files from the
previous build should still be available for comparison.  The
tests now use these files, but revert to the Table-Maker queries
if for some reason the previous build files are unavailable.